requirements.txt
fastapi
pydantic
uvicorn[standard]
python-multipart
numpy
Pillow
tensorflow
opencv-python
azure-storage-blob
google-cloud-storage